requirements.txt
pandas
ipykernel
langchain-openai
langchain-chroma
chromadb
python-dotenv